split ssi part into header and cpp file
